Refactor file structure; update import statements in tests; remove unit test due to refactoring
