+++ src/wf_psf/tests/test_data/centroids_test.py
@@ -9,7 +9,7 @@ This module contains unit tests for the wf_psf.utils centroids module.
 import numpy as np
 import pytest
 from unittest.mock import MagicMock, patch
-from wf_psf.utils.centroids import compute_zernike_tip_tilt, CentroidEstimator
+from wf_psf.data.centroids import compute_zernike_tip_tilt, CentroidEstimator
 
 
 # Function to compute centroid based on first-order moments
@@ -133,7 +133,7 @@ def test_compute_zernike_tip_tilt_single_batch(mocker, simple_image, identity_ma
     """Test compute_zernike_tip_tilt with single batch input and mocks."""
     # Mock the CentroidEstimator class
     mock_centroid_calc = mocker.patch(
-        "wf_psf.utils.centroids.CentroidEstimator", autospec=True
+        "wf_psf.data.centroids.CentroidEstimator", autospec=True
     )
 
     # Create a mock instance and configure get_intra_pixel_shifts()
@@ -144,7 +144,7 @@ def test_compute_zernike_tip_tilt_single_batch(mocker, simple_image, identity_ma
 
     # Mock shift_x_y_to_zk1_2_wavediff to return predictable values
     mock_shift_fn = mocker.patch(
-        "wf_psf.utils.centroids.shift_x_y_to_zk1_2_wavediff",
+        "wf_psf.data.centroids.shift_x_y_to_zk1_2_wavediff",
         side_effect=lambda shift: shift * 0.5,  # Mocked conversion for test
     )
 
@@ -191,7 +191,7 @@ def test_compute_zernike_tip_tilt_batch(mocker, multiple_images):
     """Test compute_zernike_tip_tilt with batch input and mocks."""
     # Mock the CentroidEstimator class
     mock_centroid_calc = mocker.patch(
-        "wf_psf.utils.centroids.CentroidEstimator", autospec=True
+        "wf_psf.data.centroids.CentroidEstimator", autospec=True
     )
 
     # Create a mock instance and configure get_intra_pixel_shifts()
@@ -202,7 +202,7 @@ def test_compute_zernike_tip_tilt_batch(mocker, multiple_images):
 
     # Mock shift_x_y_to_zk1_2_wavediff to return predictable values
     mock_shift_fn = mocker.patch(
-        "wf_psf.utils.centroids.shift_x_y_to_zk1_2_wavediff",
+        "wf_psf.data.centroids.shift_x_y_to_zk1_2_wavediff",
         side_effect=lambda shift: shift * 0.5,  # Mocked conversion for test
     )
